slå sammen filer og navnendringer
