+++ src/main/kotlin/no/nav/tms/min/side/proxy/Application.kt
@@ -1,34 +1,23 @@
 package no.nav.tms.min.side.proxy
 
-import io.ktor.client.HttpClient
-import io.ktor.server.application.port
 import io.ktor.server.engine.ApplicationEngineEnvironmentBuilder
 import io.ktor.server.engine.applicationEngineEnvironment
 import io.ktor.server.engine.connector
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
-import mu.KotlinLogging
-import no.nav.tms.min.side.proxy.config.Environment
-import no.nav.tms.min.side.proxy.config.HttpClientBuilder
-import no.nav.tms.min.side.proxy.config.mainModule
 
-private val log = KotlinLogging.logger { }
 fun main() {
-    val env = Environment()
-    val httpClient = HttpClientBuilder.build()
-
-    val envConfig = applicationEngineEnvironment { envConfig(env, httpClient) }
+    val envConfig = applicationEngineEnvironment { envConfig(AppConfiguration()) }
     embeddedServer(factory = Netty,environment = envConfig).start(wait = true)
 }
 
-fun ApplicationEngineEnvironmentBuilder.envConfig(env: Environment, httpClient: HttpClient) {
+fun ApplicationEngineEnvironmentBuilder.envConfig(appConfig: AppConfiguration) {
     rootPath = "tms-min-side-proxy"
     module {
-        mainModule(
-            corsAllowedOrigins = env.corsAllowedOrigins,
-            corsAllowedSchemes = env.corsAllowedSchemes,
-            httpClient = httpClient,
-            contentFetcher = env.contentFecther(httpClient)
+        proxyApi(
+            corsAllowedOrigins = appConfig.corsAllowedOrigins,
+            corsAllowedSchemes = appConfig.corsAllowedSchemes,
+            contentFetcher = appConfig.contentFecther
         )
     }
     connector {
