+++ src/main/kotlin/no/nav/tms/min/side/proxy/ContentFetcher.kt
@@ -1,10 +1,19 @@
 package no.nav.tms.min.side.proxy
 
 import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.request.header
+import io.ktor.client.request.request
+import io.ktor.client.request.setBody
+import io.ktor.client.request.url
 import io.ktor.client.statement.HttpResponse
+import io.ktor.http.ContentType
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpMethod
+import io.ktor.http.contentType
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import kotlinx.serialization.json.JsonElement
-import no.nav.tms.min.side.proxy.config.get
-import no.nav.tms.min.side.proxy.config.post
 import no.nav.tms.token.support.tokendings.exchange.TokendingsService
 
 class ContentFetcher(
@@ -64,4 +73,27 @@ class ContentFetcher(
         return httpClient.get(url, exchangedToken)
     }
 
+    fun shutDown() {
+        httpClient.close()
+    }
+
 }
+
+suspend inline fun <reified T> HttpClient.get(url: String, accessToken: String): T = withContext(Dispatchers.IO) {
+    request {
+        url(url)
+        method = HttpMethod.Get
+        header(HttpHeaders.Authorization, "Bearer $accessToken")
+    }.body()
+}
+
+suspend inline fun <reified T> HttpClient.post(url: String, content: JsonElement, accessToken: String): T =
+    withContext(Dispatchers.IO) {
+        request {
+            url(url)
+            method = HttpMethod.Post
+            header(HttpHeaders.Authorization, "Bearer $accessToken")
+            contentType(ContentType.Application.Json)
+            setBody(content)
+        }
+    }.body()
