--- src/main/kotlin/no/nav/tms/min/side/proxy/config/httpClient.kt
@@ -1,60 +1,0 @@
-package no.nav.tms.min.side.proxy.config
-
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.engine.apache.Apache
-import io.ktor.client.plugins.HttpTimeout
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.request.header
-import io.ktor.client.request.request
-import io.ktor.client.request.setBody
-import io.ktor.client.request.url
-import io.ktor.http.ContentType
-import io.ktor.http.HttpHeaders
-import io.ktor.http.HttpMethod
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonElement
-import java.net.URL
-
-
-object HttpClientBuilder {
-    fun build(httpClientEngine: HttpClientEngine = Apache.create()) =
-        HttpClient(httpClientEngine) {
-            install(ContentNegotiation) {
-                json(jsonConfig())
-            }
-            install(HttpTimeout)
-        }
-
-}
-
-fun jsonConfig(ignoreUnknownKeys: Boolean = false): Json {
-    return Json {
-        this.ignoreUnknownKeys = ignoreUnknownKeys
-        this.encodeDefaults = true
-    }
-}
-
-suspend inline fun <reified T> HttpClient.get(url: String, accessToken: String): T = withContext(Dispatchers.IO) {
-    request {
-        url(url)
-        method = HttpMethod.Get
-        header(HttpHeaders.Authorization, "Bearer $accessToken")
-    }.body()
-}
-
-suspend inline fun <reified T> HttpClient.post(url: String, content: JsonElement, accessToken: String): T =
-    withContext(Dispatchers.IO) {
-        request {
-            url(url)
-            method = HttpMethod.Post
-            header(HttpHeaders.Authorization, "Bearer $accessToken")
-            contentType(ContentType.Application.Json)
-            setBody(content)
-        }
-    }.body()
